# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sport Management API document",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/options": {
            "get": {
                "tags": [
                    "Option System"
                ],
                "summary": "Get options list by name list",
                "description": "Returns option list",
                "operationId": "getOptionList",
                "parameters": [
                    {
                        "name": "name[]",
                        "in": "query",
                        "description": "array name",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        },
                        "example": "['logo_name', 'logo_path']"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/api/options/{name}": {
            "get": {
                "tags": [
                    "Option System"
                ],
                "summary": "Get list of option by name",
                "description": "Returns option",
                "operationId": "getOption",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "logo",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "logo_path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/api/roles": {
            "get": {
                "tags": [
                    "Roles"
                ],
                "summary": "Get list of role by user",
                "description": "Returns role list",
                "operationId": "getRoleList",
                "parameters": [
                    {
                        "name": "type",
                        "in": "query",
                        "description": "[All/Role]",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "All"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Roles"
                ],
                "summary": "create role",
                "description": "create role",
                "operationId": "createRole",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Loai nguoi dung moi",
                                    "description": "mo ta"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Roles"
                ],
                "summary": "delete role",
                "description": "delete role",
                "operationId": "deleteRole",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Roles"
                ],
                "summary": "update role",
                "description": "update role",
                "operationId": "updateRole",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Loai nguoi dung moi",
                                    "description": "mo ta"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
...
    }
}